ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.25.2
openai==1.3.7
whisper==1.1.10
torch==2.1.0
torchaudio==2.1.0
silero-vad==4.0.0
coqui-tts==0.20.0
numpy==1.24.3
scipy==1.11.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

